refactor: 필요 없는 커스텀 훅 제거 및 이름 변경
